% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_select.R
\name{compare_select}
\alias{compare_select}
\title{Compare feature selection methods}
\usage{
compare_select(dataset, resVar, varNum)
}
\arguments{
\item{dataset}{A dataframe.}

\item{resVar}{A string with response variable name.}

\item{varNum}{A number of features to select in random forest method.}
}
\value{
A list with results table and selected variables for each method.
}
\description{
This function combines results of 3 selection methods on given data.
}
\examples{
res <- compare_select(mtcars, "hp", 3)
}
