% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_select.R
\name{rf_select}
\alias{rf_select}
\title{Select variables with random forest method}
\usage{
rf_select(responseVar, varNum, train, test)
}
\arguments{
\item{responseVar}{A string corresponding to response variable name.}

\item{varNum}{Number of features to be selected.}

\item{train}{A train subset of data.}

\item{test}{A test subset of data.}
}
\value{
A list with rmse of random forest model and vector of selected variables.
}
\description{
This function builds random forest model, select n most important variables and calculates rmse.
}
\examples{
df <- mtcars
split <- caTools::sample.split(df, SplitRatio = 0.7)
train <- df[split, ]
test <- df[!split, ]
res <- rf_select("hp", 3, train, test)
}
