% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{show_results}
\alias{show_results}
\title{Show results of feature selection comparison on data from package}
\usage{
show_results()
}
\description{
This is driver function for project.
}
\examples{
show_results()
}
