% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select.R
\name{step_select}
\alias{step_select}
\title{Select variables with step algorithm}
\usage{
step_select(responseVar, train, test)
}
\arguments{
\item{responseVar}{A string corresponding to response variable name.}

\item{train}{A train subset of data.}

\item{test}{A test subset of data.}
}
\value{
A list containing rmse and variables of full and step models.
}
\description{
This function builds full and step models of given data and calculate rmse.
}
\examples{
df <- mtcars
split <- caTools::sample.split(df, SplitRatio = 0.7)
train <- df[split, ]
test <- df[!split, ]
res <- step_select("hp", train, test)
}
