% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Quality Data Set}
\format{
A data frame with 1599 rows and 12 variables:
\describe{
\item{fixed.acidity}{numeric}
\item{volatile.acidity}{numeric}
\item{citric.acid}{numeric}
\item{residual.sugar}{numeric}
\item{chlorides}{numeric}
\item{free.sulfur.dioxide}{numeric}
\item{total.sulfur.dioxide}{numeric}
\item{density}{numeric}
\item{pH}{numeric}
\item{sulphates}{numeric}
\item{alcohol}{numeric}
\item{quality}{int score between 0 and 10}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Wine+Quality}
}
\usage{
wine
}
\description{
The dataset related to red variant of the Portuguese "Vinho Verde" wine.
}
\keyword{datasets}
